gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.24.1
pydantic==2.4.2
PyYAML==6.0.1 